Added photo for future lessons and enhanced Main.kt with boolean variables and time formatting function
In this commit, I added a photo that will be used in future lessons. Additionally, I included new useful code in the Main.kt file to demonstrate the usage of boolean variables. I also implemented a helpful function for time formatting. Furthermore, I created directories to organize content for future lessons. These additions improve the overall functionality and structure of the project, setting the foundation for upcoming lessons.