Add function to check number parity and demo code
Added a new function to check whether a given number is even or odd. The function implements a simple logic to determine parity based on the remainder when dividing the number by 2. Additionally, included code to demonstrate the usage of the function by prompting the user to enter a number and displaying the result.