Added code for funny number guessing game on Main.kt and new lessons 😄
Added code to the Main.kt file that implements a cool number guessing game. The code allows the user to input the range of numbers, generates a random number within that range, and prompts the user to guess the generated number. The user is given a limited number of attempts to guess the number correctly. The code also calculates the user's score based on the range difficulty and the number of remaining attempts.

Additionally, included 8 new lessons and associated images. 🎉